+++ R/GUI_MANOVA.R
@@ -10,7 +10,7 @@
 #' data needs to be provided.
 #' For the resampling methods, the user can choose between a parametric bootstrap approach 
 #' (see e.g. Konietschke et al. (2015)) and a Wild bootstrap using Rademacher weights
-#' (see e.g. Bathke et al. (2016)).
+#' (see e.g. Bathke et al. (2018)).
 #' 
 #' 
 #' @export